backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-environ==0.11.2
python-decouple==3.8

# Database
dj-database-url==2.1.0
psycopg2-binary==2.9.7

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
dj-rest-auth==4.0.1

# Internationalization
django-modeltranslation==0.18.12
django-parler==2.4

# API Documentation
drf-spectacular==0.26.5

# Image Processing
Pillow==10.0.1

# Development
django-debug-toolbar==4.2.0

# Production
whitenoise==6.6.0
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0

# PDF Generation
reportlab==4.0.7 
arabic-reshaper==3.0.0
python-bidi==0.4.2